### 2019-05-25 #1475 @danielskatz
👋 @trallard - you've been suggested as editor - is this ok with you?
