### 2019-05-25 9174bd9
Fix author name in paper.md
