### 2019-05-28 942f765
Start making pulse compatible with fenics version 2018.
