### 2019-06-02 419c3df
All tests pass in serial expect biv elliosoid - suspect that something is broken in HDF5Attributes
