+++ pulse/material/material_model.py
@@ -143,7 +143,7 @@ class Material(object):
                     v_new = RegionalParameter(geometry.sfun)
                     numpy_mpi.assign_to_vector(
                         v_new.vector(),
-                        numpy_mpi.gather_broadcast(v.vector().get_local()),
+                        numpy_mpi.gather_vector(v.vector()),
                     )
                     v = v_new
 
