### 2019-06-17 #1475 @melund
Sorry for missing this @trallard. I can take the review :)
