### 2019-07-01 #1475 @trallard
Hi fellas I do not want to hold this any further so since this aligns well with my areas of expertise I am going to self assign me as a reviewer and start the process
