### 2019-07-01 #1539 @whedon
[REVIEW]: pulse
**Submitting author:** @finsberg (<a href="http://orcid.org/0000-0003-3766-2393">Henrik Nicolay Topnes Finsberg</a>)
**Repository:** <a href="https://github.com/finsberg/pulse" target ="_blank">https://github.com/finsberg/pulse</a>
**Version:** 2019.0

- [x] **Version:** 2019.0

- [x] **Version:** 2019.0
