### 2019-07-01 #1539 @melund
Thanks, @trallard. I may be a little while before I can get to this. I am going on sommer holiday on Friday. And I will most likely be away from my computer for at least two weeks.

I am sorry, but I am just too busy before I leave to get this done. I will look at it when I get back :)
