### 2019-07-24 8cbd41a
Update joss paper
