+++ joss/paper.md
@@ -4,8 +4,8 @@ tags:
   - Python
   - FEniCS
   - Cardiac Mechanics
-  - Finite Element
-  - Mechanics
+  - Finite Element Method
+  - Continuum Mechanics
 authors:
   - name: Henrik Nicolay Topnes Finsberg
     orcid: 0000-0003-3766-2393
@@ -15,29 +15,23 @@ affiliations:
  - name: Simula Research Laboratory, Oslo, Norway
    index: 1
 
-date: 24 May 2019
+date: 24 July 2019
 bibliography: paper.bib
 ---
 # Summary
 
-`pulse` is a software based on FEniCS [@logg2012automated] that can
-solve general problems in continuum mechanics but was designed to be
-used to model the mechanics of the heart at the organ level.
-At this level, the heart, or can be modeled as a hyperelastic and
-incompressible continuum body embedded in three dimesional Euclidean
-space. However, the user is free to define whatever material model and
-impressibility conditions he/she like. 
+`pulse` is a software based on [FEniCS](https://fenicsproject.org) [@logg2012automated] that aims to solve problems in cardiac mechanics (but is easily extended to solve more general problems in continuum mechanics). `pulse` is a results of the author's PhD thesis [@finsberg2017patient], where most of the relevant background for the code can be found.
 
-As a part of the code it there is a quite sophisticated solver that
-uses numerical continuation for stepping up the parameters.
+ While FEniCS offers a general framework for solving PDEs, `pulse` specifically targets problems in continuum mechanics. Therefore, most of the code for applying compatible boundary conditions, formulating the governing equations, choosing appropriate spaces for the solutions and applying iterative strategies etc. are already implemented, so that the user can focus on the actual problem he/she wants to solve rather than implementing all the necessary code for formulating and solving the underlying equations. 
 
-It is also implemented an algorithm, known as the backward
-displacement method for finding the unloaded stress free geometry.
+The user can pick any of the built-in meshes or choose a custom user defined mesh. The user also need to provide appropriate markers for the boundaries where the boundary conditions will be applied, as well as microstructural information (i.e information about muscle fiber orientations) if an anisotropic model is to be used. Examples of how to create custom idealized geometries as well as appropriate microstructure can be found in another repository called [ldrb](https://github.com/finsberg/ldrb) which uses the Laplace-Dirichlet Rule-Based (LDRB) algorithm for assigning myocardial fiber orientations.
 
-For a detail background in cardiac mechanics and the methods used in
-this code the reader is referred to [@finsberg2017patient].
+Next the user needs to select a material model or create a [custom material model](https://finsberg.github.io/pulse/html/demos/custom_material.html), and define appropriate boundary conditions (Dirichlet, Neumann or Robin boundary conditions). Finally a `MechanicsProblem` is built using the geometry, material and boundary conditions. Figure ![Visualization of the different components that are part of the `pulse` library](components.png) shows the different components involved as well as how they are related. The problem is solved using some iterative strategy, either with an incremental load technique with fixed or adaptive stepping and/or using with a continuation technique [@finsberg2017patient].
 
-Papers using this code includes [@finsberg2018estimating] and
-[@finsberg2018efficient].
+It is also possible to estimate the unloaded zero-pressure geometry [@bols2013computational]. This is of particular importance if the geometry you are using is taken from a medical image of a patient. In this case, the geometry is subjected to some load due to the blood pressure, and therefore in order to correctly assess the stresses one need to first first the unloaded geometry.
+
+Papers using this code includes [@finsberg2018estimating] and [@finsberg2018efficient].
+
[…]
