### 2019-07-24 05ee924
Update REAMDE and docs as suggested by @melund
