+++ README.md
@@ -6,12 +6,12 @@
 
 
 A software for solving problems in cardiac mechanics.
-The code in this repo used to be part of [pulse-adjoint](https://bitbucket.org/finsberg/pulse_adjoint), but now works as a standalone mechanics solver without the need for dolfin-adjoint.
+The code in this repository used to be part of [pulse-adjoint](https://bitbucket.org/finsberg/pulse_adjoint), but now works as a standalone mechanics solver without the need for dolfin-adjoint.
 
 ## Overview
 `pulse` is a software based on [FEniCS](https://fenicsproject.org) that aims to solve problems in cardiac mechanics (but is easily extended to solve more general problems in continuum mechanics). `pulse` is a results of the author's [PhD thesis](https://www.duo.uio.no/handle/10852/62015), where most of the relevant background for the code can be found.
 
- While FEniCS offers a general framework for solving PDEs, `pulse` specifically targets problems in continuum mechanics. Therefore, most of the code for applying compatible boundary conditions, formulating the governing equations, choosing appropriate spaces for the solutions and applying iterative strategies etc. are already implemented, so that the user can focus on the actual problem he/she wants to solve rather than implementing all the necessary code for formulating and solving the underlying equations. 
+While FEniCS offers a general framework for solving PDEs, `pulse` specifically targets problems in continuum mechanics. Therefore, most of the code for applying compatible boundary conditions, formulating the governing equations, choosing appropriate spaces for the solutions and applying iterative strategies etc. are already implemented, so that the user can focus on the actual problem he/she wants to solve rather than implementing all the necessary code for formulating and solving the underlying equations. 
 
 ## Installation instructions
 
@@ -42,9 +42,9 @@ docker pull finsberg/pulse:latest
 ```
 
 ## Requirements
-* FEniCS verision 2017.x
+* FEniCS version 2017.x
 
-Note that if you install FEniCS using anaconda then you will not get support for paralell HDF5
+Note that if you install FEniCS using anaconda then you will not get support for parallel HDF5
 see e.g [this issue](https://github.com/conda-forge/hdf5-feedstock/issues/51).
 
 ## Getting started
@@ -59,8 +59,8 @@ python -m pytest tests -vv
 
 
 ## Documentation
-Documentation can be found at[finsberg.github.io/pulse](https://finsberg.github.io/pulse)
-You can create documentation youselves by typing `make html` in the
+Documentation can be found at [finsberg.github.io/pulse](https://finsberg.github.io/pulse)
+You can create documentation yourselves by typing `make html` in the
 root directory.
 
 
