### 2019-08-01 762177c
Fix compatibility issue in HDF5Attributes
