### 2019-08-01 f4b96b3
Update docker image to be fenics latest
