+++ .circleci/config.yml
@@ -8,7 +8,7 @@ jobs:
     docker:
       # specify the version you desire here
       # use `-browsers` prefix for selenium tests, e.g. `3.6.1-browsers`
-      - image: quay.io/fenicsproject/stable:2017.2.0
+      - image: quay.io/fenicsproject/stable:latest
       
       # Specify service dependencies here if necessary
       # CircleCI maintains a library of pre-built images
