### 2019-08-02 ed4751a
Fix mpi_comm_world typo in demo
