+++ demo/from_xml/demo.py
@@ -5,7 +5,7 @@ import json
 
 
 # Mesh
-mesh = dolfin.Mesh(dolfin.mpi_comm_world(), "data/mesh.xml")
+mesh = dolfin.Mesh(pulse.utils.mpi_comm_world(), "data/mesh.xml")
 
 # Marker functions
 facet_function = dolfin.MeshFunction("size_t", mesh, "data/facet_function.xml")
@@ -21,10 +21,9 @@ for k, v in markers_dict.items():
     markers.append(marker)
 
 # Fiber
-fiber_element = dolfin.VectorElement(family="Quadrature",
-                                     cell=mesh.ufl_cell(),
-                                     degree=4,
-                                     quad_scheme="default")
+fiber_element = dolfin.VectorElement(
+    family="Quadrature", cell=mesh.ufl_cell(), degree=4, quad_scheme="default"
+)
 fiber_space = dolfin.FunctionSpace(mesh, fiber_element)
 fiber = dolfin.Function(fiber_space, "data/fiber.xml")
 
@@ -32,37 +31,40 @@ microstructure = pulse.Microstructure(f0=fiber)
 
 
 # Create the geometry
-geometry = pulse.HeartGeometry(mesh=mesh, markers=markers,
-                               marker_functions=marker_functions,
-                               microstructure=microstructure)
+geometry = pulse.HeartGeometry(
+    mesh=mesh,
+    markers=markers,
+    marker_functions=marker_functions,
+    microstructure=microstructure,
+)
 
 activation = dolfin.Function(dolfin.FunctionSpace(geometry.mesh, "R", 0))
 activation.assign(dolfin.Constant(0.2))
 matparams = pulse.HolzapfelOgden.default_parameters()
-material = pulse.HolzapfelOgden(activation=activation,
-                                parameters=matparams,
-                                f0=geometry.f0)
+material = pulse.HolzapfelOgden(
+    activation=activation, parameters=matparams, f0=geometry.f0
+)
 
 # LV Pressure
 lvp = dolfin.Constant(1.0)
-lv_marker = markers_dict['ENDO'][0]
-lv_pressure = pulse.NeumannBC(traction=lvp,
-                              marker=lv_marker, name='lv')
+lv_marker = markers_dict["ENDO"][0]
+lv_pressure = pulse.NeumannBC(traction=lvp, marker=lv_marker, name="lv")
 neumann_bc = [lv_pressure]
 
 # Add spring term at the base with stiffness 1.0 kPa/cm^2
 base_spring = 1.0
-robin_bc = [pulse.RobinBC(value=dolfin.Constant(base_spring),
-                          marker=markers_dict["BASE"][0])]
+robin_bc = [
+    pulse.RobinBC(value=dolfin.Constant(base_spring), marker=markers_dict["BASE"][0])
+]
 
 
 # Fix the basal plane in the longitudinal direction
 # 0 in V.sub(0) refers to x-direction, which is the longitudinal direction
 def fix_basal_plane(W):
     V = W if W.sub(0).num_sub_spaces() == 0 else W.sub(0)
-    bc = dolfin.DirichletBC(V.sub(0),
-                            dolfin.Constant(0.0),
-                            geometry.ffun, markers_dict["BASE"][0])
+    bc = dolfin.DirichletBC(
+        V.sub(0), dolfin.Constant(0.0), geometry.ffun, markers_dict["BASE"][0]
+    )
     return bc
 
 
@@ -74,9 +76,9 @@ dirichlet_bc = [fix_basal_plane]
 #                        marker=geometry.markers["BASE"][0])
 
[…]
