### 2019-08-02 8b3727f
Some fixes for parallel execution
