+++ setup.py
@@ -12,20 +12,20 @@ from shutil import rmtree
 from setuptools import find_packages, setup, Command
 
 # Package meta-data.
-NAME = 'fenics-pulse'
-DESCRIPTION = ('A python library based on FEniCS that aims to solve '
-               'problems in continuum mechanics, in particular '
-               'cardiac mechanics.')
-URL = 'https://github.com/finsberg/pulse'
-EMAIL = 'henriknf@simula.no'
-AUTHOR = 'Henrik Finsberg'
-REQUIRES_PYTHON = '>=3.5.0'
-VERSION = '0.1.4'
+NAME = "fenics-pulse"
+DESCRIPTION = (
+    "A python library based on FEniCS that aims to solve "
+    "problems in continuum mechanics, in particular "
+    "cardiac mechanics."
+)
+URL = "https://github.com/finsberg/pulse"
+EMAIL = "henriknf@simula.no"
+AUTHOR = "Henrik Finsberg"
+REQUIRES_PYTHON = ">=3.5.0"
+VERSION = "2019.0"
 
 # What packages are required for this module to be executed?
-REQUIRED = [
-    'h5py', 'numpy', 'scipy'
-]
+REQUIRED = ["h5py", "numpy", "scipy"]
 
 # What packages are optional?
 EXTRAS = {
@@ -42,8 +42,8 @@ here = os.path.abspath(os.path.dirname(__file__))
 # Import the README and use it as the long-description.
 # Note: this will only work if 'README.md' is present in your MANIFEST.in file!
 try:
-    with io.open(os.path.join(here, 'README.md'), encoding='utf-8') as f:
-        long_description = '\n' + f.read()
+    with io.open(os.path.join(here, "README.md"), encoding="utf-8") as f:
+        long_description = "\n" + f.read()
 except FileNotFoundError:
     long_description = DESCRIPTION
 
@@ -51,22 +51,22 @@ except FileNotFoundError:
 about = {}
 if not VERSION:
     project_slug = NAME.lower().replace("-", "_").replace(" ", "_")
-    with open(os.path.join(here, project_slug, '__version__.py')) as f:
+    with open(os.path.join(here, project_slug, "__version__.py")) as f:
         exec(f.read(), about)
 else:
-    about['__version__'] = VERSION
+    about["__version__"] = VERSION
 
 
 class UploadCommand(Command):
     """Support setup.py upload."""
 
-    description = 'Build and publish the package.'
+    description = "Build and publish the package."
     user_options = []
 
     @staticmethod
     def status(s):
         """Prints things in bold."""
-        print('\033[1m{0}\033[0m'.format(s))
+        print("\033[1m{0}\033[0m".format(s))
 
     def initialize_options(self):
         pass
@@ -76,31 +76,31 @@ class UploadCommand(Command):
 
     def run(self):
         try:
-            self.status('Removing previous builds…')
-            rmtree(os.path.join(here, 'dist'))
+            self.status("Removing previous builds…")
+            rmtree(os.path.join(here, "dist"))
         except OSError:
             pass
 
-        self.status('Building Source and Wheel (universal) distribution…')
-        os.system('{0} setup.py sdist bdist_wheel --universal'.format(sys.executable))
+        self.status("Building Source and Wheel (universal) distribution…")
+        os.system("{0} setup.py sdist bdist_wheel --universal".format(sys.executable))
 
-        self.status('Uploading the package to PyPI via Twine…')
-        os.system('python -m twine upload dist/*')
+        self.status("Uploading the package to PyPI via Twine…")
+        os.system("python -m twine upload dist/*")
 
-        self.status('Pushing git tags…')
-        os.system('git tag v{0}'.format(about['__version__']))
[…]
