### 2019-08-02 9165771
Update README and conda dependencies
