+++ conda.recipe/meta.yaml
@@ -16,59 +16,9 @@ build:
   noarch: python
 
 requirements:
-  build:
-    - python 3.6.4=0
   run:
     - h5py>=2.8.0
-    - blas=1.1 openblas
-    - boost 1.65.1
-    - boost-cpp 1.65.1=1
-    - bzip2 1.0.6
-    - ca-certificates 2017.11.5=0
-    - certifi 2017.11.5=py36_0
-    - cmake 3.10.0=1
-    - eigen 3.3.3=0
-    - expat 2.2.5=0
-    - fastcache 1.0.2=py36_0
-    - gmp 6.1.2=0
-    - gmpy2 2.0.8=py36_1
-    - hdf5 1.10.1=1
-    - icu 58.2=0
-    - krb5 1.14.2=0
-    - libgfortran 3.0.1
-    - libssh2 1.8.0=2
-    - libuv 1.11.0=0
-    - metis 5.1.0=3
-    - mpc 1.0.3=4
-    - mpfr 3.1.5=0
-    - mpmath 1.0.0=py_0
-    - ncurses 5.9=10
-    - numpy 1.13.3=py36_blas_openblas_200
-    - openblas 0.2.19=2
-    - openssl 1.0.2n=0
-    - pcre 8.39=0
-    - petsc 3.8.1=blas_openblas_0
-    - petsc4py 3.8.1=py36_1
-    - pip 9.0.1=py36_1
-    - pkg-config 0.29.1=2
-    - ply 3.10=py36_0
-    - ptscotch 6.0.4=4
-    - readline 7.0=0
-    - rhash 1.3.4=0
-    - scotch 6.0.4=3
-    - setuptools 38.2.4=py36_0
-    - six 1.11.0=py36_1
-    - slepc 3.8.0=blas_openblas_0
-    - slepc4py 3.8.0=py36_0
-    - sqlite 3.20.1=2
-    - suitesparse 4.5.4=blas_openblas_200
-    - swig 3.0.12=2
-    - sympy 1.1.1=py36_0
-    - tbb 2018_20170919=0
-    - tk 8.6.7=0
-    - xz 5.2.3=0
-    - zlib 1.2.11=0
-    - fenics 2017.2
+    - fenics 2018
 test:
   imports:
     - pulse
