+++ environment.yml
@@ -3,89 +3,7 @@ channels:
   - conda-forge
   - defaults
 dependencies:
-  - atomicwrites=1.3.0=py_0
-  - attrs=18.2.0=py_0
-  - blas=1.1=openblas
-  - boost=1.65.1=py36_0
-  - boost-cpp=1.65.1=1
-  - bzip2=1.0.6=1
-  - ca-certificates=2019.3.9=hecc5488_0
-  - certifi=2019.3.9=py36_0
-  - cmake=3.10.0=1
-  - curl=7.60.0=0
-  - cycler=0.10.0=py_1
-  - eigen=3.3.3=0
-  - expat=2.2.5=0
-  - fastcache=1.0.2=py36_0
-  - fenics=2017.2.0=np113py36_0
-  - freetype
-  - gmp=6.1.2=0
-  - gmpy2=2.0.8=py36_1
-  - h5py=2.8.0=py36h470a237_0
-  - hdf5=1.10.1=1
-  - icu=58.2=0
-  - kiwisolver
-  - krb5=1.14.2=0
-  - libcxx=8.0.0=2
-  - libcxxabi=8.0.0=2
-  - libgfortran
-  - libpng
-  - libssh2=1.8.0=2
-  - libuv=1.11.0=0
-  - llvm-meta=7.0.0=0
-  - matplotlib
-  - metis=5.1.0=3
-  - more-itertools=4.3.0=py36_1000
-  - mpc=1.0.3=4
-  - mpfr=3.1.5=0
-  - mpi=1.0=mpich
-  - mpich
-  - mpmath=1.0.0=py_0
-  - mshr=2017.2.0=py36_0
-  - ncurses=5.9=10
-  - numpy=1.13.3=py36_blas_openblas_200
-  - openblas=0.2.19=2
-  - openssl
-  - pcre=8.39=0
-  - petsc=3.8.1=blas_openblas_0
-  - petsc4py=3.8.1=py36_1
-  - pip=9.0.1=py36_1
-  - pkg-config=0.29.1=2
-  - pluggy=0.9.0=py_0
-  - ply=3.10=py36_0
-  - ptscotch=6.0.4=4
-  - py=1.8.0=py_0
-  - pyparsing=2.3.1=py_0
-  - pytest=3.9.1=py36_1000
-  - python=3.6.4=0
-  - python-dateutil=2.8.0=py_0
-  - readline=7.0=0
-  - rhash=1.3.4=0
-  - scotch=6.0.4=3
-  - setuptools=38.2.4=py36_0
-  - six=1.11.0=py36_1
-  - slepc=3.8.0=blas_openblas_0
-  - slepc4py=3.8.0=py36_0
-  - sqlite=3.20.1=2
-  - suitesparse=4.5.4=blas_openblas_200
-  - swig=3.0.12=2
-  - sympy=1.1.1=py36_0
-  - tbb=2018_20170919=0
-  - tk
-  - tornado
-  - wheel=0.33.1=py36_0
-  - xz=5.2.3=0
-  - zlib=1.2.11=0
-  - pip:
-    - fenics-dijitso==2017.2.0
-    - fenics-fiat==2017.2.0
-    - fenics-instant==2017.2.0
-    - fenics-pulse
-    - fenics-ufl==2017.2.0
-    - instant==2017.1.0
-    - ldrb
-    - llvmlite==0.28.0
-    - numba
-    - numpy-quaternion
-    - scipy==1.3.0
-
+  - fenics=2018
+  - h5py
+  - scipy
+  - numpy
