### 2019-08-02 4cf495a
Update conda recipe
