### 2019-08-05 6ee1c53
Update docs with conda-forge installation
