+++ demo/simple_ellipsoid/lv_demo.py
@@ -1,9 +1,14 @@
 import matplotlib.pyplot as plt
 import dolfin
 import pulse
+import sys
 
+if len(sys.argv) > 1:
+    gamma_space = sys.argv[1]
+else:
+    gamma_space = "R_0"
 
-geometry = pulse.Geometry.from_file(pulse.mesh_paths['simple_ellipsoid'])
+geometry = pulse.Geometry.from_file(pulse.mesh_paths["simple_ellipsoid"])
 # Plot geometry
 # dolfin.plot(geometry.mesh, edgecolor="k", color="w")
 # ax = plt.gca()
@@ -11,35 +16,45 @@ geometry = pulse.Geometry.from_file(pulse.mesh_paths['simple_ellipsoid'])
 # ax.set_axis_off()
 # plt.show()
 
-activation = dolfin.Function(dolfin.FunctionSpace(geometry.mesh, "R", 0))
-activation.assign(dolfin.Constant(0.2))
+if gamma_space == "regional":
+    activation = pulse.RegionalParameter(geometry.cfun)
+    target_activation = pulse.dolfin_utils.get_constant(0.2, len(activation))
+else:
+    activation = dolfin.Function(dolfin.FunctionSpace(geometry.mesh, "R", 0))
+    target_activation = dolfin.Constant(0.2)
+
+activation.assign(target_activation)
 matparams = pulse.HolzapfelOgden.default_parameters()
-material = pulse.HolzapfelOgden(activation=activation,
-                                parameters=matparams,
-                                f0=geometry.f0,
-                                s0=geometry.s0,
-                                n0=geometry.n0)
+material = pulse.HolzapfelOgden(
+    activation=activation,
+    parameters=matparams,
+    f0=geometry.f0,
+    s0=geometry.s0,
+    n0=geometry.n0,
+)
 
 # LV Pressure
 lvp = dolfin.Constant(1.0)
-lv_marker = geometry.markers['ENDO'][0]
-lv_pressure = pulse.NeumannBC(traction=lvp,
-                              marker=lv_marker, name='lv')
+lv_marker = geometry.markers["ENDO"][0]
+lv_pressure = pulse.NeumannBC(traction=lvp, marker=lv_marker, name="lv")
 neumann_bc = [lv_pressure]
 
 # Add spring term at the base with stiffness 1.0 kPa/cm^2
 base_spring = 1.0
-robin_bc = [pulse.RobinBC(value=dolfin.Constant(base_spring),
-                          marker=geometry.markers["BASE"][0])]
+robin_bc = [
+    pulse.RobinBC(
+        value=dolfin.Constant(base_spring), marker=geometry.markers["BASE"][0]
+    )
+]
 
 
 # Fix the basal plane in the longitudinal direction
 # 0 in V.sub(0) refers to x-direction, which is the longitudinal direction
 def fix_basal_plane(W):
     V = W if W.sub(0).num_sub_spaces() == 0 else W.sub(0)
-    bc = dolfin.DirichletBC(V.sub(0),
-                            dolfin.Constant(0.0),
-                            geometry.ffun, geometry.markers["BASE"][0])
+    bc = dolfin.DirichletBC(
+        V.sub(0), dolfin.Constant(0.0), geometry.ffun, geometry.markers["BASE"][0]
+    )
     return bc
 
 
@@ -51,9 +66,9 @@ dirichlet_bc = [fix_basal_plane]
 #                        marker=geometry.markers["BASE"][0])
 
 # Collect boundary conditions
-bcs = pulse.BoundaryConditions(dirichlet=dirichlet_bc,
[…]
