### 2019-08-05 #1539 @melund
I am done with my review. I think pulse more than qualifies for JOSS. It is a really good example of how to structure an open-source software project. With the recent addition of conda-forge package, I am also sure that the library will continue to work in the future.

All in all, nice work @finsberg  👍
