### 2019-08-16 2eaafc9
Update environment.yml
