+++ Dockerfile
@@ -1,6 +1,0 @@
-FROM finsberg/fenics2017
-MAINTAINER Henrik Finsberg <henriknf@simula.no>
-
-RUN sudo pip3 install git+https://github.com/finsberg/pulse.git
-RUN sudo pip3 install git+https://github.com/finsberg/ldrb.git
-
