### 2019-08-16 867e054
Add pulse from conda-forge to environment.yml in order to enable binder
