+++ environment.yml
@@ -8,3 +8,4 @@ dependencies:
     - fenics
     - scipy
     - matplotlib
+    - pulse
