### 2019-08-25 #1539 @trallard
Thanks @melund
I just noticed that the item for the license has not been checked on your list but I just double checked and this is in compliance. So @finsberg  I am going to move to the following editorial activites.

Can you please make sure to perform the following actions:
1. Generate a new DOI for the software and past it here
2. Make sure that the package version is up to date and report here: if needed make a new release and report it here
3. Do a last review of the paper
