### 2019-08-25 aef7b8f
Update joss paper + add doi in bibtex file
