+++ joss/paper.bib
@@ -14,6 +14,7 @@
   number={10},
   pages={2243--2254},
   year={2012},
+  doi={10.1007/s10439-012-0593-5},
   publisher={Springer}
 }
 
@@ -24,6 +25,7 @@
   volume={24},
   pages={85--90},
   year={2018},
+  doi={10.1016/j.jocs.2017.07.013 },
   publisher={Elsevier}
 }
 
@@ -35,6 +37,7 @@
   number={7},
   pages={e2982},
   year={2018},
+  doi={10.1002/cnm.2982}, 
   publisher={Wiley Online Library}
 }
 
@@ -53,5 +56,18 @@
   volume={246},
   pages={10--17},
   year={2013},
+  doi={10.1016/j.cam.2012.10.034},
   publisher={Elsevier}
+}
+
+@article{land2015verification,
+  title={Verification of cardiac mechanics software: benchmark problems and solutions for testing active and passive material behaviour},
+  author={Land, Sander and Gurev, Viatcheslav and Arens, Sander and Augustin, Christoph M and Baron, Lukas and Blake, Robert and Bradley, Chris and Castro, Sebastian and Crozier, Andrew and Favino, Marco and others},
+  journal={Proceedings of the Royal Society A: Mathematical, Physical and Engineering Sciences},
+  volume={471},
+  number={2184},
+  pages={20150641},
+  year={2015},
+  doi={10.1098/rspa.2015.0641},
+  publisher={The Royal Society Publishing}
 }
