### 2019-08-25 #1539 @finsberg
1. @trallard How do you want me do generate a DOI for the software? [Zenodo](https://zenodo.org)?
2. Package version number [2019.0](https://github.com/finsberg/pulse/releases/tag/v2019.0) is up to date.
3. I have done final review of the paper, and made a few corrections. The last article proof looks good. I also added the DOIs in the bib-file.
