### 2019-08-26 #1539 @melund
@trallard I forgot to check the license box. Everything is fine now. Good work @finsberg
