### 2019-09-04 3c2bac1
Avoid TypeError when supplying 'bcs = None' if bcs_parameters is also None
