+++ pulse/mechanicsproblem.py
@@ -157,6 +157,8 @@ class MechanicsProblem(object):
         if bcs is None:
             if isinstance(geometry, HeartGeometry):
                 self.bcs_parameters = MechanicsProblem.default_bcs_parameters()
+                if bcs_parameters is not None:
+                    self.bcs_parameters.update(**bcs_parameters)
                 self.bcs_parameters.update(**bcs_parameters)
 
             else:
