### 2019-09-24 #1539 @trallard
@finsberg  I have completed my editorial checks and I am willing to recommend this for publication 🎉

@melund thanks for your contribution as a reviewer 🙌🏼

@openjournals/jose-eics  I am passing this submission to you for acceptance
