### 2019-09-24 #1539 @danielskatz
👋 @trallard - fwiw, the accepted label will be automatically attached during the final processing
