### 2019-09-24 #1539 @trallard
Thanks for the heads up. I have been following the instructions at https://joss.readthedocs.io/en/latest/editing.html#after-acceptance where it says the editor should add the `accepted` label
