### 2019-09-24 66b7b3c
minor edits
