### 2019-09-24 #1539 @danielskatz
👋 @finsberg - please merge in the changes in https://github.com/finsberg/pulse/pull/8 and https://github.com/finsberg/pulse/pull/9 or let me know what you disagree with
